fix moodtrackerfragment

+++ app/src/main/java/com/example/frontend_moodify/data/remote/response/journal/Moods.kt
@@ -1,9 +1,9 @@
 package com.example.frontend_moodify.data.remote.response.journal
 
 data class Moods(
-    val anger: Int,
-    val enthusiasm: Int,
-    val happiness: Int,
-    val sadness: Int,
-    val worry: Int
+    val anger: Float,
+    val enthusiasm: Float,
+    val happiness: Float,
+    val sadness: Float,
+    val worry: Float
 )
